Support .svg files
This commit relies on GNU sed; as Firefox renders some .svg graphics
incorrectly except when the width is nudged to be a little bit wider,
we need to inspect every .svg file to determine the width, and add
a bit to that.

To this end, upload.sh uses GNU sed's "e" option to the s/// command:
execute the command given by the pattern space.

The command is simply upload.sh with "handle" as first parameter, and
the match as second parameter.

It is a bit of forking, but avoids the use of Perl.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>