Introduce identifiers for fsck messages
Rather than specifying only whether a message by the fsck machinery
constitutes an error or a warning, let's specify an identifier relating
to the concrete problem that was encountered. This is necessary for
upcoming support to be able to demote certain errors to warnings.

While at it, simplify the requirements on the calling code: instead of
having to handle full-blown varargs, we now send a string buffer ready to
be used by the callback.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>