Avoid gcc compiler warning
At least on this developer's MacOSX (Snow Leopard, gcc-4.2.1), GCC prints
a warning that 'hash' may be used uninitialized when compiling
test-hashmap that 'hash' may be used uninitialized (but GCC 4.6.3 on this
developer's Ubuntu server does not report this problem).

The old compiler is wrong, of course, as the switch(method & 3) statement
already handles all the possible cases, but that does not help in a
scenario where it is hard or impossible to upgrade to a newer compiler
(e.g. being stuck on an older MacOSX and having to rely on Xcode).

So let's just initialize the variable and be done with it, it is hardly a
crucial part of the code because it is only used by the test suite and
invisible to the end users.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>