fsck: Allow demoting errors to warnings
For example, missing emails in commit and tag objects can be demoted to
mere warnings with

	git config receive.fsck.severity missing-email=warn

The value is actually a comma-separated list.

In case that the same key is listed in multiple receive.fsck.severity
lines in the config, the latter configuration wins (this can happen for
example when both $HOME/.gitconfig and .git/config contain severity
settings).

As git receive-pack does not actually perform the checks, it hands off
the setting to index-pack or unpack-objects in the form of an optional
argument to the --strict option.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>