WIP: Add a test for non-boolean receive.fsckobjects values
TODO!!! This only *happens* to work because the fsck machinery is
ill-prepared to handle warnings. It is based on the concept that we can
simply error out when there is so much as even a warning, but what we want
to do is to try to figure things out and continue even after detecting
that, say, an email is missing or invalid.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>