Avoid accessing uninitialized alt_odb entry data
At the end of link_alt_odb_entry, we replace the trailing NUL of the base
directory by a slash. The idea is that the users will want to append
either the "pack/" directory or the path to a loose object.

However, at least `git prune` wants to access the base directory itself.
Therefore, we need to NUL-terminate after appending a slash to the base
directory, to avoid undeterministic behavior.

Thanks "Valerie" Valgrind!

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>