When re-executing tests, remember when they were called via `sh -x`
When debugging tests, it is a good practice to call them via `sh -x
t0123-something.sh -i -v`. The `-x` option asks the shell to print out the
commands before executing them.

For some modes, however, we have to re-execute the shell, e.g. when
running with the --valgrind flag. Let's not forget to call the subshell
with the same `-x` flag.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>