log: teach --invert-grep option
"git log --grep=<string>" shows only commits with messages that
match the given string, but sometimes it is useful to be able to
show only commits that do *not* have certain messages (e.g. "show
me ones that are not FIXUP commits").

Introduce a new --invert-grep option, and add a corresponding bit in
the rev_info struct that tells us "we'd run the usual grep thing on
the log message, and instead of filtering to show only the commits
with matching message, show the ones with messages that do not
match".

Signed-off-by: Christoph Junghans <ottxor@gentoo.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>