mingw: fix git-svn tests that expect chmod to work
Some git-svn tests expect that the executable bit of files can be
toggled. On Windows, this is not possible because Windows' Access
Control Lists are much more fine-grained than the POSIX permission
concept. Let's just not try to flip the executable bit.

This lets t9118-git-svn-funky-branch-names.sh and
t9124-git-svn-dcommit-auto-props.sh pass in Git for Windows' SDK,
and partially fixes t9100-git-svn-basic.sh.

Signed-off-by: 마누엘 <nalla@users.noreply.github.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>