engine.pl: name the msvc buildsystem's makedry error file
Save the stderr from the dry MSVC make to a well named file for
later review. Use 'msvc-build-makedryerrors.txt' which should be
obvious as to its source, and is not ignored by 'git status'.

Signed-off-by: Philip Oakley <philipoakley@iee.org>
---
fixed false mode change:- Junio's report (last line of)
http://marc.info/?l=git&m=143524504500824 (2015-06-25),

The cause was probably an early 'commit --amend' during a
rebase, and using the git gui to unstage the whole commit, then
pick up and commit hunks one at a time. Unfortunately, on Msysgit/g4w
core.filemode is false, so the execute bit was lost during the
unstage, never noticed, and there's no simple way of adding it back.

Discussion on correcting file mode on msysgit/git4Windows is at
https://groups.google.com/forum/#!topic/msysgit/zwH-qj0xR48