merge-file: consider core.crlf when writing merge markers
When merging files in repos with core.eol = crlf, git merge-file inserts
just a LF at the end of the merge markers. Files with mixed line endings
cause trouble in Windows editors and e.g. contrib/git-jump, where an
unmerged file in a run of "git jump merge" is reported as simply "binary
file matches".

Fixing this improves merge-file's behavior under Windows.

The original version of this patch also modified ll_merge(), but that
was incorrect: low-level merge operates on blobs, not on working files.
Therefore, the data passed to the low-level merge, as well as its
result, is expected to have LF-only line endings.

It is the duty of ll_merge()'s *caller* (in case of Git's `merge`
command, the merge_content() function) to convert the merge result into
the correct working file contents, and ll_merge() should not muck with
line endings at all.

Signed-off-by: Beat Bolli <dev+git@drbeat.li>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>