hack: watchman/untracked cache mashup
DO NOT APPLY THIS

This is a hack to add untracked cache info to the watchman WAMA
extension.  The idea is that we don't have to stat directories if
watchman hasn't told us of any changes in those dirs.

If this is the right idea, I can go back and add it to the relevant
patch(es).  But I want to see if other folks think it's the right idea
first.

Signed-off-by: David Turner <dturner@twopensource.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>